implemented sending emails for friendship requests
